Parse cluster! input into ItemStructs for flexibility and profit
