+++ example/Cargo.toml
@@ -7,5 +7,5 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ni_fpga = { path = "../ni_fpga" }
-ni_fpga_macros = { path = "../ni_fpga_macros" }
+ni-fpga = { path = "../ni-fpga" }
+ni-fpga-macros = { path = "../ni-fpga-macros" }
