+++ ni-fpga-derive/Cargo.toml
@@ -1,12 +1,14 @@
 [package]
-name = "ni_fpga_macros"
+name = "ni-fpga-derive"
 version = "0.1.0"
 authors = ["Connor Worley <connorbworley@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+proc-macro = true
+
 [dependencies]
-ni_fpga_derive = { path = "../ni_fpga_derive" }
-packed_struct = "0.3.0"
-packed_struct_codegen = "0.3.0"
+quote = "1.0.2"
+syn = { version = "1.0.11", features = ["full"] }
