+++ ni-fpga-macros/Cargo.toml
@@ -1,14 +1,12 @@
 [package]
-name = "ni_fpga_derive"
+name = "ni-fpga-macros"
 version = "0.1.0"
 authors = ["Connor Worley <connorbworley@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[lib]
-proc-macro = true
-
 [dependencies]
-quote = "1.0.2"
-syn = { version = "1.0.11", features = ["full"] }
+ni-fpga-derive = { path = "../ni-fpga-derive" }
+packed_struct = "0.3.0"
+packed_struct_codegen = "0.3.0"
